requirements.txt
APScheduler~=3.10.1
fastapi~=0.94.1
jose~=1.0.0
loguru~=0.6.0
peewee~=3.16.0
starlette~=0.26.1
uvicorn~=0.21.0
pymysql
passlib[bcrypt]
pydantic[dotenv]
python-jose[cryptography]
redis~=5.0.1
websockets~=12.0
httpx~=0.25.0
PyJWT~=2.8.0
requests
kubernetes==28.1.0
casdoor==1.17.0
python-pam==2.0.2
#python-ldap==3.4.4